metaheuristics/search stuff/requirements.txt
termcolor
python-dotenv